# Improve sessions management in health tests