--- a/src/engine/test/health_test/engine-health-test/src/health_test/test_suite.py
+++ b/src/engine/test/health_test/engine-health-test/src/health_test/test_suite.py
@@ -6,7 +6,7 @@ from pathlib import Path
 from typing import List, Tuple, Optional, Union
 from abc import ABC, abstractmethod
 from engine_handler.handler import EngineHandler
-
+from shared.default_settings import Constants
 
 class UnitOutput:
     def __init__(self, index: int, result: Union[str, dict]):
@@ -188,7 +188,7 @@ def run_test(test_parent_path: Path, engine_api_socket: str, unit_result: type, 
 
         ns = "wazuh system" if target == 'rule' else "wazuh"
         engine_test_command = f"engine-test -c {engine_test_conf.resolve().as_posix()} "
-        engine_test_command += f"run {test_name} --api-socket {engine_api_socket} -n {ns} {debug_mode} -j"
+        engine_test_command += f"run {test_name} -s {Constants.DEFAULT_SESSION} --api-socket {engine_api_socket} -n {ns} {debug_mode} -j"
         command = f"cat {input_file.resolve().as_posix()} | {engine_test_command}"
         test_result = test(input_file, expected_file,
                            unit_result, command, target, help)