|Related issue|
|---|
|#29318|


## Description

This PR addresses performance bottlenecks in dynamic E2E testing by implementing session reuse within the health-test suite. Previously, each test case triggered a new invocation of engine-test, resulting in redundant session initializations and significantly slower runs.

Changes Introduced
- Session Initialization: A single session is now created at the beginning of the health-test run.

- Session Reuse: The same session is passed to each engine-test invocation throughout the suite.

- Graceful Cleanup: The session is properly closed after all tests finish, with handling for failures to ensure cleanup still occurs.

- Benchmarking: Preliminary benchmarks show a noticeable reduction in execution time for dynamic E2E tests.


## Impact
By avoiding repeated session creation, this change leads to faster test execution, improving CI throughput and reducing local iteration times for developers.