@NahuFigueroa97 2025-04-21
## Comparison of execution times

### Old implementation
```console
(venv) ╭─root@404dfb28aa28 /workspaces/devContainer/wazuh/src/engine ‹main●› 
╰─# time engine-health-test dynamic -e /tmp/prueba run --integration suricata
Validating environment...
Environment validated.
Starting engine...
Specified integration: suricata
2025-04-21 19:23:07.194 27485:27485 info: Logging initialized.
2025-04-21 19:23:07.195 27485:27485 info: Skipping configuration from API.
Engine started.


Running tests...
....Engine stopped.



suricata -> Success
All tests passed.
engine-health-test dynamic -e /tmp/prueba run --integration suricata  78.37s user 0.98s system 99% cpu 1:19.61 total
```

### New implementation
```console
(venv) ╭─root@404dfb28aa28 /workspaces/devContainer/wazuh/src/engine ‹enhancement/29318-optimize-health-tests●› 
╰─# time engine-health-test dynamic -e /tmp/prueba run --integration suricata                                               
Validating environment...
Environment validated.
Starting engine...
Specified integration: suricata
2025-04-21 19:28:40.952 30235:30235 info: Logging initialized.
2025-04-21 19:28:40.953 30235:30235 info: Skipping configuration from API.
Engine started.


Running tests...
....Engine stopped.



suricata -> Success
All tests passed.
engine-health-test dynamic -e /tmp/prueba run --integration suricata  10.41s user 0.39s system 86% cpu 12.462 total
```