@NahuFigueroa97 2025-04-21
Old execution
https://github.com/wazuh/intelligence-data/actions/runs/14402858892/job/40392335993

New execution
https://github.com/wazuh/intelligence-data/actions/runs/14581485774/job/40898975883